requirements.txt
netmiko==4.4.0
PyYAML==6.0
